old_version/requirements.txt
pysha3
py_ecc
git+https://github.com/ethereum/pyrlp
